+++ packages/components-native/src/Menu/index.ts
@@ -1,0 +1,6 @@
+export { Menu } from "./Menu";
+export type {
+  MenuProps,
+  MenuOptionProps,
+  MenuOptionInternalProps,
+} from "./types";